Modularised the code and made the main script a bit more clean
